feat: replace some shipping views with unified savedAddresses
