+++ packages/theme/components/MyAccount/SavedAddressForm.vue
@@ -1,7 +1,7 @@
 <template>
   <ValidationObserver v-slot="{ handleSubmit }">
     <form
-      id="shipping-details-form"
+      id="savedAddresses-details-form"
       class="form"
       @submit.prevent="handleSubmit(submitForm)"
     >
@@ -12,7 +12,7 @@
           class="form__element"
         >
           <SfInput
-            data-cy="shipping-details-input_firstName"
+            data-cy="savedAddresses-details-input_firstName"
             v-model="form.firstName"
             name="firstName"
             label="First Name"
@@ -27,7 +27,7 @@
           class="form__element"
         >
           <SfInput
-            data-cy="shipping-details-input_lastName"
+            data-cy="savedAddresses-details-input_lastName"
             v-model="form.lastName"
             name="lastName"
             label="Last Name"
@@ -43,7 +43,7 @@
         class="form__element"
       >
         <SfInput
-          data-cy="shipping-details-input_streetName"
+          data-cy="savedAddresses-details-input_streetName"
           v-model="form.addressLine1"
           name="streetName"
           label="Street Name"
@@ -53,7 +53,7 @@
         />
       </ValidationProvider>
       <SfInput
-        data-cy="shipping-details-input_apartment"
+        data-cy="savedAddresses-details-input_apartment"
         v-model="form.addressLine2"
         name="apartment"
         label="House/Apartment number"
@@ -67,7 +67,7 @@
           class="form__element"
         >
           <SfInput
-            data-cy="shipping-details-input_city"
+            data-cy="savedAddresses-details-input_city"
             v-model="form.city"
             name="city"
             label="City"
@@ -83,7 +83,7 @@
           class="form__element"
         >
           <SfSelect
-            data-cy="shipping-details-input_state"
+            data-cy="savedAddresses-details-input_state"
             class="form__select sf-select--underlined"
             v-model="form.state"
             name="state"
@@ -109,7 +109,7 @@
           class="form__element"
         >
           <SfInput
-            data-cy="shipping-details-input_zipCode"
+            data-cy="savedAddresses-details-input_zipCode"
             v-model="form.postalCode"
             name="zipCode"
             label="Zip-code"
@@ -125,7 +125,7 @@
           class="form__element"
         >
           <SfSelect
-            data-cy="shipping-details-select_country"
+            data-cy="savedAddresses-details-select_country"
             class="form__select sf-select--underlined"
             v-model="form.country"
             name="country"
@@ -150,7 +150,7 @@
         class="form__element"
       >
         <SfInput
-          data-cy="shipping-details-input_phoneNumber"
+          data-cy="savedAddresses-details-input_phoneNumber"
           v-model="form.phone"
           name="phone"
           label="Phone number"
